requirements.txt
matplotlib
numpy
opencv-python
Pillow
scikit-learn
scipy
torch
torchvision
tqdm 
pandas
torchsummary
tensorboardX